requirements.txt
torch
torchvision
torchmetrics
numpy
pandas
opencv-python
scipy
scikit-learn
seaborn
hydra-core
timm
python-dotenv 
wandb 
tqdm
pillow